pyproject.toml
[project]
name = "hr-policy-rag-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "chromadb>=1.5.7",
    "docx2txt>=0.9",
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-huggingface>=1.2.1",
    "langchian>=0.2.5",
    "numpy>=2.4.4",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.26",
    "ragas>=0.4.3",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.4.0",
    "streamlit>=1.56.0",
]
